refactor: extract CodeEditorBase from CodeEditor
Splits the CodeEditor control into an abstract CodeEditorBase that owns the
host, and a sealed CodeEditor that is just the plain-editor flavor. This is
preparation for a sibling DiffCodeEditor control: the lifecycle machinery in
CodeEditorBase.Events.cs carries three independent init-recovery paths that
must not be duplicated.

Behaviorally a no-op. The four CodeEditor partials become CodeEditorBase
partials; CodeEditor keeps only its constructors, DefaultStyleKey, and the
Text dependency property.

Text stays on the concrete control rather than the base so that neither
flavor carries an overloaded name -- DiffCodeEditor will expose OriginalText
and ModifiedText instead. The base reaches the value through a new
PrimaryText hook.

New extension points on the base, all of them the delta a derived control
needs: BootstrapFunctionName, IsDiffEditor, PrimaryText,
BuildInitialStateJson, ApplyInitialPropertyValues, and
RegisterBridgeCallbacks. BuildCreateMonacoEditorScript gains an overload
taking the bootstrap function name; the existing single-argument overload is
retained so CodeEditorInitSequenceTests keeps testing it unchanged.

Retypes CodeEditor -> CodeEditorBase on the surfaces that reference the
control: ICodeEditorPresenter.ParentCodeEditor, IActionDescriptor.Run,
WebKeyEventHandler, and the ModelHelper/LanguagesHelper/CssStyleBroker/
ContextKey constructors. Nothing has shipped under the Uno.Monaco.Editor
package id yet, so these are not breaking changes in practice.
