+++ MonacoEditorComponent/Monaco/LanguagesHelper.cs
@@ -13,14 +13,14 @@ namespace Monaco
     /// completion, hover, code-action, code-lens, and color providers.
     /// </summary>
     /// <remarks>
-    /// Obtain an instance from <see cref="CodeEditor.Languages"/>. Do not construct directly.
+    /// Obtain an instance from <see cref="CodeEditorBase.Languages"/>. Do not construct directly.
     /// See <see href="https://microsoft.github.io/monaco-editor/typedoc/modules/editor_editor_api.languages.html">monaco.languages</see>.
     /// </remarks>
     [method: Obsolete("Use <Editor Instance>.Languages.* instead of constructing your own LanguagesHelper.")]
     [method: EditorBrowsable(EditorBrowsableState.Never)]
-    public sealed partial class LanguagesHelper(CodeEditor editor)
+    public sealed partial class LanguagesHelper(CodeEditorBase editor)
     {
-        private readonly WeakReference<CodeEditor> _editor = new(editor);
+        private readonly WeakReference<CodeEditorBase> _editor = new(editor);
 
         /// <summary>
         /// Gets the list of registered language identifiers and their extension points.
