+++ MonacoEditorTestApp/EditorControl.xaml.cs
@@ -93,7 +93,7 @@ namespace MonacoEditorTestApp
             Debug.WriteLine("Property changed - " + e.PropertyName);
         }
 
-        private void Editor_InternalException(CodeEditor sender, Exception args)
+        private void Editor_InternalException(CodeEditorBase sender, Exception args)
         {
             // This shouldn't happen, if it does, then it's a bug.
         }
@@ -587,7 +587,7 @@ namespace MonacoEditorTestApp
                 .Replace("\r", "\\r", StringComparison.Ordinal)
                 .Replace("\n", "\\n", StringComparison.Ordinal)}\"";
 
-        private void Editor_OpenLinkRequest(CodeEditor sender, OpenLinkRequestedEventArgs args)
+        private void Editor_OpenLinkRequest(CodeEditorBase sender, OpenLinkRequestedEventArgs args)
         {
             if (this.AllowWeb.IsChecked == false)
             {
@@ -663,7 +663,7 @@ namespace MonacoEditorTestApp
         }
 
         // Note: Can't make this method async as otherwise handled won't be read for intercepts.
-        private void Editor_KeyDown(CodeEditor sender, WebKeyEventArgs e)
+        private void Editor_KeyDown(CodeEditorBase sender, WebKeyEventArgs e)
         {
             Debug.WriteLine("KeyDown: " + e.KeyCode + " " + e.CtrlKey);
